Refactor architecture violation handling and introduce new data classes
