+++ popcornguineapigplugin/src/main/kotlin/com/github/codandotv/popcorn/domain/models/ArchitectureViolationError.kt
@@ -1,6 +1,5 @@
-package com.github.codandotv.popcorn.domain.output
+package com.github.codandotv.popcorn.domain.models
 
-import com.github.codandotv.popcorn.domain.metadata.InternalDependenciesMetadata
 import com.github.codandotv.popcorn.domain.rules.PopcornGuineaPigRule
 
 public data class ArchitectureViolationError(
