+++ popcornguineapigplugin/src/main/kotlin/com/github/codandotv/popcorn/domain/models/CheckResult.kt
@@ -1,7 +1,5 @@
 package com.github.codandotv.popcorn.domain.models
 
-import com.github.codandotv.popcorn.domain.output.ArchitectureViolationError
-
 internal sealed class CheckResult {
     object Success : CheckResult()
