+++ popcornguineapigplugin/src/main/kotlin/com/github/codandotv/popcorn/domain/models/TargetModule.kt
@@ -1,6 +1,5 @@
 package com.github.codandotv.popcorn.domain.models
 
-import com.github.codandotv.popcorn.domain.metadata.InternalDependenciesMetadata
 import com.github.codandotv.popcorn.domain.rules.PopcornGuineaPigRule
 
 internal data class TargetModule(
