+++ popcornguineapigplugin/src/main/kotlin/com/github/codandotv/popcorn/domain/rules/PopcornGuineaPigRule.kt
@@ -1,7 +1,7 @@
 package com.github.codandotv.popcorn.domain.rules
 
-import com.github.codandotv.popcorn.domain.output.ArchitectureViolationError
-import com.github.codandotv.popcorn.domain.metadata.InternalDependenciesMetadata
+import com.github.codandotv.popcorn.domain.models.ArchitectureViolationError
+import com.github.codandotv.popcorn.domain.models.InternalDependenciesMetadata
 
 public interface PopcornGuineaPigRule {
