+++ popcornguineapigplugin/src/test/kotlin/com/github/codandotv/popcorn/domain/models/TargetModuleTest.kt
@@ -1,6 +1,5 @@
-package com.github.codandotv.popcorn.domain.metadata
+package com.github.codandotv.popcorn.domain.models
 
-import com.github.codandotv.popcorn.domain.models.TargetModule
 import kotlin.test.Test
 import kotlin.test.assertEquals
