+++ popcornguineapigplugin/src/test/kotlin/com/github/codandotv/popcorn/domain/usecases/GenerateReportUseCaseTest.kt
@@ -1,6 +1,6 @@
 package com.github.codandotv.popcorn.domain.usecases
 
-import com.github.codandotv.popcorn.domain.output.ArchitectureViolationError
+import com.github.codandotv.popcorn.domain.models.ArchitectureViolationError
 import com.github.codandotv.popcorn.domain.models.CheckResult
 import com.github.codandotv.popcorn.domain.rules.NoDependencyRule
 import com.github.codandotv.popcorn.fakes.fakePopcornGuineapigRepository
